+++ examples/database/flex_component/main.tf
@@ -1,0 +1,29 @@
+// Copyright (c) 2017, 2020, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+variable "tenancy_ocid" {}
+variable "user_ocid" {}
+variable "fingerprint" {}
+variable "private_key_path" {}
+variable "region" {}
+variable "compartment_ocid" {}
+
+variable "flex_component_name" {
+  default = "name"
+}
+
+provider "oci" {
+  tenancy_ocid     = var.tenancy_ocid
+  user_ocid        = var.user_ocid
+  fingerprint      = var.fingerprint
+  private_key_path = var.private_key_path
+  region           = var.region
+}
+
+data "oci_database_flex_components" "test_flex_components" {
+  #Required
+  compartment_id = "${var.compartment_ocid}"
+
+  #Optional
+  name = var.flex_component_name
+}